added initial version of preferences window
not sure how to deal with locale. i believe the best choice is to change the default one, but still unsure on how to proceed
Opted for a Save method instead of a ChangeListener since it'd be pretty useless
GridPane seems to be a good fit! Also included a css file to try that out. Seems fun