Redefine WeightFunction and related classes
This revision tackles the difficulty of transmitting edge weight changes to interested
observing classes. The task was accomplished with the aid of Java FX binding and
properties. Instead of having a WeightFunction inherit from Observable (and thus be monitored
for changes) EdgeWeight classes do. So in short, WeigthFunctions instantiate weights for
a graph, and they in turn can transmit state changes to those instances which are interested
in receiving them.