Initial version of fixing one very very very annoying todo
UNFINISHED! works for Lessons, not for Questions. Wanted to share the code and show I've been banging my head on this for a while.
fileSystem both in LessonLoader and QuestionLoader need to be closed, but I am unsure as to when that should be done. Maybe as the window is closed?
I am 100% sure you have a preferred way of handling exceptions, and I am not sure what the best one is. This is the only way I have found to fix this issue, and it is based on https://stackoverflow.com/a/28057735
If you have time, this other one points out security issues but seems excessive
https://stackoverflow.com/questions/1429172/how-to-list-the-files-inside-a-jar-file/37730853#37730853
This was a useful check of understanding of matchers and regular expression! Unfortunately most of the other stuff went above my head. If you understand this code perfectly, I'll probably have a couple questions for you